+++ src/parser/ast_lexer/mod.rs
@@ -28,7 +28,7 @@ pub(crate) struct AstLexer<'a> {
 /// (for example 555.55 could be matched by both float and integer (integer can just match the first
 /// part of it)) so it's important float is first. Another example is comments - they have to be
 /// stripped out first
-fn matchers_ordered(tl: &TokenLibrary) -> [&TokenMatcher<Token>; 17] {
+fn matchers_ordered(tl: &TokenLibrary) -> [&TokenMatcher<Token>; 16] {
     [
         &tl.newline,
         &tl.comment,
@@ -39,8 +39,7 @@ fn matchers_ordered(tl: &TokenLibrary) -> [&TokenMatcher<Token>; 17] {
         &tl.comma,
         &tl.close_paren,
         &tl.open_paren,
-        &tl.infix_operator,
-        &tl.postfix_operator,
+        &tl.operator,
         &tl.code_section_eof,
         // float has to be happen before integer!  it needs to greedy match 1.5, where integer will
         // also match the first part 1, but not the rest
@@ -193,8 +192,8 @@ mod tests {
         assert_token_match_eq!(lexer, Token::Comma, ",", 1, 2);
         assert_token_match_eq!(lexer, Token::Reference, "b", 1, 4);
         assert_token_match_eq!(lexer, Token::CloseParen, ")", 1, 5);
-        assert_token_match_eq!(lexer, Token::InfixOperator, "+", 1, 7);
-        assert_token_match_eq!(lexer, Token::InfixOperator, "*", 1, 9);
+        assert_token_match_eq!(lexer, Token::Operator, "+", 1, 7);
+        assert_token_match_eq!(lexer, Token::Operator, "*", 1, 9);
         assert_token_match_eq!(lexer, Token::Float, "0.25", 1, 11);
         assert_token_match_eq!(lexer, Token::Eof, "", 1, 15);
         assert_token_match_eq!(lexer, Token::Eof, "", 1, 15);